Created classes to forward activites properly
