+++ ftc-cameratest/build.gradle
@@ -28,12 +28,3 @@ dependencies {
     compile project(':opencv-java')
     compile project(':ftc-visionlib')
 }
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-    compile 'com.journeyapps:zxing-android-embedded:3.0.3@aar'
-    compile 'com.google.zxing:core:3.2.0'
-}
