+++ ftc-visionlib/build.gradle
@@ -38,3 +38,12 @@ dependencies {
 
     compile 'org.jetbrains:annotations:13.0'
 }
+
+repositories {
+    jcenter()
+}
+
+dependencies {
+    compile 'com.journeyapps:zxing-android-embedded:3.0.3@aar'
+    compile 'com.google.zxing:core:3.2.0'
+}
