+++ ftc-visionlib/src/main/java/org/lasarobotics/vision/test/detection/QRDetector.java
@@ -1,0 +1,20 @@
+package org.lasarobotics.vision.test.detection;
+
+import android.app.Activity;
+
+import com.google.zxing.integration.android.IntentIntegrator;
+
+/**
+ * Created by Russell on 11/26/2015.
+ */
+public class QRDetector {
+    Activity context;
+
+    public QRDetector(Activity context) {
+        this.context = context;
+    }
+
+    public void intiateDetection() {
+        //IntentIntegrator.
+    }
+}
